requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
typer==0.9.0
rich==13.7.0

# Database and storage
sqlalchemy==2.0.23
sqlite3  # Built-in with Python

# Natural Language Processing
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4

# Text similarity and embeddings
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Configuration and utilities
pyyaml==6.0.1
python-dateutil==2.8.2

# Development dependencies
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional: For future web interface
fastapi==0.104.1
uvicorn==0.24.0
